pyproject.toml
[tool.poetry]
name = "yoko"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google = "^3.0.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.1"
flux = {git = "https://github.com/black-forest-labs/flux.git"}
fal-client = "^0.4.1"
rich = "^13.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
